refactor(cloudflare-metrics): split metrics.ts by concern
the 318-line metrics.ts mixed four unrelated concerns — the Metric data
structure, the retry-buffer + last-flush state, the two provider
implementations, and the CloudflareMetricsRepository facade. split into
four files:

- metric.ts                 Metric data class
- flush-state.ts            module-level retry buffer + last-flush stats
                            (takeLastFlushStats / recordLastFlushStats)
- metric-providers.ts       HeaderMetricsProvider + InfluxMetricsProvider
- metrics.ts                CloudflareMetricsRepository + re-exports

metrics.ts keeps re-exports so downstream imports don't move yet; a
follow-up commit will switch call sites to the canonical paths.
